[podannotation] Move AddRoutesGatewayIP to the allocator/pod.
This helps to avoid circular dependency with generator/udn
